Move interactive map js into subfolder
