+++ src/aims_ui/static/js/a_single_matches/radiussearch/init.mjs
@@ -1,5 +1,5 @@
 import { setupSubmissionFromInputs } from "./form_helpers.mjs";
-import { updateMapToFitCircle, setupLatLongToMapListeners, setSizeOfMapToPreviouslySetSize, setupResizeListeners, addMatchedAddressMarkersToMap, setupMap, setupRadiusListeners, setupLatLongFromMapListeners, setupInitialMarkerLocation } from "./interactive_map.mjs";
+import { updateMapToFitCircle, setupLatLongToMapListeners, setSizeOfMapToPreviouslySetSize, setupResizeListeners, addMatchedAddressMarkersToMap, setupMap, setupRadiusListeners, setupLatLongFromMapListeners, setupInitialMarkerLocation } from "./interactive_map/interactive_map.mjs";
 import { setupHorizontalInputs } from "./page_reformatting.mjs";
 import { setupScrollListeners, setScrollPosition } from "./scroll_helpers.mjs";
 import { setupUprnSearchFunctionality } from "./uprn_helper.mjs";
