+++ src/aims_ui/static/js/a_single_matches/radiussearch/interactive_map/interactive_map_icons.mjs
